# faster A_mul_B! and * involving bidiagonal and tridiagonal matrices