Two `A_mul_B!(C, A, B)` methods where `B` or `A, B` are tridiagonal are implemented. This gives faster matrix multiplication methods for some bi and tridiagonal argument combinations, which would fall back to full matrix multiplication before. See #13358 and timings below.

```
begin
           import Base.LinAlg.A_mul_B!
           k = 1000;
           A = Tridiagonal(randn(k-1), randn(k), randn(k-1));
           B = Tridiagonal(randn(k-1), randn(k), randn(k-1));
           C = randn(k, k);

           Ai = Bidiagonal(randn(k), randn(k-1), rand(Bool));
           Bi = Bidiagonal(randn(k), randn(k-1), rand(Bool));
           fullAi = full(Ai);
           fullBi = full(Bi);
           Atri = Tridiagonal(randn(k-1), randn(k), randn(k-1));
           Btri = Tridiagonal(randn(k-1), randn(k), randn(k-1));
           C = randn(k, k);
       end
```

```
julia> @time A_mul_B!(C, fullAi, fullBi);
  0.035776 seconds (4 allocations: 160 bytes)

julia> @time A_mul_B!(C, fullAi, Btri);
  0.002102 seconds (4 allocations: 160 bytes)

julia> @time A_mul_B!(C, Atri, fullBi);
  0.004037 seconds (4 allocations: 160 bytes)

julia> @time A_mul_B!(C, Atri, Btri);
  0.000815 seconds (4 allocations: 160 bytes)
```
```
julia> @time A_mul_B!(C, fullAi, Bi);
  0.001518 seconds (11 allocations: 16.141 KB)

julia> @time A_mul_B!(C, Ai, fullBi);
  0.004141 seconds (11 allocations: 16.141 KB)

julia> @time A_mul_B!(C, Ai, Bi);
  0.001324 seconds (18 allocations: 32.125 KB)
```
```
julia> @time fullAi * fullBi;
  0.035390 seconds (9 allocations: 7.630 MB)

julia> @time fullAi * Btri;
  0.001393 seconds (9 allocations: 7.630 MB)

julia> @time Atri * fullBi;
  0.004084 seconds (9 allocations: 7.630 MB)

julia> @time Atri * Btri;
  0.001264 seconds (9 allocations: 7.630 MB)
```
```
julia> @time fullAi * Bi;
  0.002728 seconds (16 allocations: 7.645 MB)

julia> @time Ai * fullBi;
  0.005010 seconds (16 allocations: 7.645 MB)

julia> @time Ai * Bi;
  0.001789 seconds (23 allocations: 7.661 MB)

```