@andreasnoack 2016-03-15
Thanks. It's great with the specialized methods.