@jw3126 2016-03-15
Thanks for all the feedback. Btw have you any idea, why the AppVeyor build failed?