@pkofod 2016-03-15
> Thanks for all the feedback. Btw have you any idea, why the AppVeyor build failed?

Someone might have a sharper eye than I do, but I think you just lucked out. The AppVeyor plan used allows for the total run time to be no more than 60 minutes. The entire thing on `ARCH=x86_64` usually runs at around 52-53 minutes, so I think you were just unlucky.