@jw3126 2016-05-01
Rebased and squashed, hopefully ready to merge now.