@jw3126 2016-05-04
@tkelman Because life is too short to be waiting for sparse arrays being filled by inefficient methods.
```
julia> N = 10^6
1000000

julia> sparr = spzeros(N);

julia> arr = zeros(N);

julia> @time fill!(arr, 2);
  0.001341 seconds (4 allocations: 160 bytes)

julia> @time fill!(sparr, 2);
  0.004634 seconds (6 allocations: 15.259 MB)
```