@tkelman 2016-05-12
I wish this were achievable in a bit less code, but since it probably isn't, I think this is ready. Couple minor style nits.