@jw3126 2016-05-12
I agree there are some not so elegant places in this code, but I don't see how to make it shorter. Though I think writing a code generator, which creates A_mul_B! methods such as these automatically for lots of sparsity + memory layout patterns is in principle possible. Of course that would be a much much bigger project.

Big thanks @tkelman and the others for all that feedback!